Refactor vehicle model validation and input types for numeric fields

+++ database/migrations/2025_02_25_154524_create_vehicle_models_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->string('manufacturer')->nullable()->comment('vehicle manufacturer / brand');
-            $table->integer('engine_cc')->nullable();
+            $table->double('engine_cc')->nullable();
             $table->double('fuel_capacity')->nullable()->comment('max fuel capacity in liters');
             $table->double('payload_capacity')->nullable()->comment('max load capacity in kg');
             $table->double('body_length')->nullable()->comment('in feet');
